requirements.txt
# requirements.txt
# Список всех Python библиотек необходимых для работы системы

Flask==2.3.3          # 🌐 Микро-фреймворк для веб-сервера
requests==2.31.0       # 📡 Для HTTP запросов к Figma API
python-dotenv==1.0.0   # 🔑 Для работы с .env файлами (токены и ключи)
gunicorn==21.2.0       # 🚀 Production веб-сервер для Flask (для деплоя)

# Flask - создает API endpoints для взаимодействия с системой
# requests - отправляет запросы к Figma API для получения данных о дизайне  
# python-dotenv - безопасно загружает секретные ключи из .env файла
# gunicorn - запускает Flask приложение в production среде